// wizard-library/foundation.d.ts
export type WizardInputElement = SclTextfield | TextField | TextArea | Select | SclSelect;
/** @returns [[`EditorAction`]]s to dispatch on [[`WizardDialog`]] commit. */
export type WizardActor = (inputs: WizardInputElement[], wizard: Element, list?: List | null) => Edit[];
// ...
export interface MenuAction {
    label: string;
    icon?: string;
    action: WizardMenuActor;
}
/** Represents a page of a wizard dialog */
export interface WizardPage {
    title: string;
    content?: TemplateResult[];
    primary?: {
        icon: string;
        label: string;
        action: WizardActor;
        auto?: boolean;
    };
    secondary?: {
        icon: string;
        label: string;
        action: WizardActor;
    };
    initial?: boolean;
    menuActions?: MenuAction[];
}
export type Wizard = WizardPage[];